+++ wandb/sdk/wandb_settings.py
@@ -356,6 +356,7 @@ class SettingsData:
     _python: str
     _runqueue_item_id: str
     _require_core: bool
+    _require_legacy_service: bool
     _save_requirements: bool
     _service_transport: str
     _service_wait: float
@@ -650,7 +651,7 @@ class Settings(SettingsData):
             },
             _disable_service={
                 "value": False,
-                "preprocessor": _str_as_bool,
+                "preprocessor": self._process_disable_service,
                 "is_policy": True,
             },
             _disable_setproctitle={"value": False, "preprocessor": _str_as_bool},
@@ -718,6 +719,7 @@ class Settings(SettingsData):
                 "preprocessor": _str_as_json,
             },
             _require_core={"value": False, "preprocessor": _str_as_bool},
+            _require_legacy_service={"value": False, "preprocessor": _str_as_bool},
             _save_requirements={"value": True, "preprocessor": _str_as_bool},
             _service_wait={
                 "value": 30,
@@ -1171,6 +1173,16 @@ class Settings(SettingsData):
             raise UsageError("hostname is invalid")
 
         return True
+
+    @staticmethod
+    def _process_disable_service(value: Union[str, bool]) -> bool:
+        value = _str_as_bool(value)
+        if value:
+            wandb.termwarn(
+                "Disabling the wandb service is deprecated as of version 0.18.0 and will be removed in version 0.19.0.",
+                repeat=False,
+            )
+        return value
 
     @staticmethod
     def _validate__service_wait(value: float) -> bool: